move character component to character.rs
