+++ src/components/spaceship.rs
@@ -7,18 +7,6 @@ use amethyst::{
     ecs::prelude::{Component, DenseVecStorage},
 };
 
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct CharacterComponent {
-    pub money: usize,
-    pub collision_damage: f32,
-}
-
-impl Component for CharacterComponent {
-    type Storage = DenseVecStorage<Self>;
-}
-
 pub struct SpaceshipComponent {
     pub barrel_cooldown: f32,
     pub barrel_reset_timer: f32,
